requirements.txt
numpy
scipy
matplotlib
dill
pandas
torch==1.6.0
tensorboard==2.5.0
tensorflow-gpu==1.15.0
advertorch
